fix readme task imgs 2

+++ readme/render_github_document.R
@@ -1,8 +1,16 @@
+debug <- FALSE
+
 file.move <- function(from, to, overwrite = TRUE) {
   if (file.copy(from, to, overwrite))
     unlink(from)
 }
 
 rmarkdown::render("README.Rmd", rmarkdown::github_document())
-unlink("README.html")
+
+if (debug) {
+  file.move("README.html", "../README.html")
+} else {
+  unlink("README.html")
+}
+
 file.move("README.md", "../README.md")
